font-metric module
