+++ rendering/renderer.js
@@ -1,5 +1,6 @@
 import { promise, num } from "lively.lang";
 import { addOrChangeCSSDeclaration, addOrChangeLinkedCSS } from "./dom-helper.js";
+import FontMetric from "./font-metric.js";
 import { defaultStyle, defaultAttributes, render } from "./morphic-default.js";
 import {h} from "virtual-dom";
 
@@ -95,14 +96,13 @@ export class Renderer {
       domEnvironment = {window: System.global, document: doc};
     }
     this.worldMorph = world;
-    world._isWorld = true; // for world() method
     world._renderer = this;
     this.rootNode = rootNode;
     this.domNode = null;
     this.domEnvironment = domEnvironment;
     this.renderMap = new WeakMap();
     this.renderWorldLoopProcess = null;
-    FontMetric.initDefault(domEnvironment.document);
+    FontMetric.initDefault({document: domEnvironment.document});
   }
 
   clear() {
@@ -279,114 +279,4 @@ function renderGradient(morph) {
                               attributes:
                                 {offset: (k * 100) + "%",
                                  "stop-color": c}}))));
-}
-
-
-export class FontMetric {
-
-  static default() {
-    if (!this._fontMetric)
-      throw new Error("FontMetric has not yet been initialized!")
-    return this._fontMetric;
-  }
-
-  static initDefault(doc = typeof document !== "undefined" ? document : null) {
-    if (!this._fontMetric) {
-      this._fontMetric = new FontMetric();
-      this._fontMetric.install(doc, doc.body);
-    }
-  }
-
-  static removeDefault() {
-    if (this._fontMetric) {
-      this._fontMetric.uninstall();
-      this._fontMetric = null;
-    }
-  }
-
-  constructor() {
-    this.charMap = [];
-    this.parentElement = null;
-    this.element = null;
-  }
-
-  install(doc, parentEl) {
-    this.parentElement = parentEl;
-    this.element = doc.createElement("div");
-    this.setMeasureNodeStyles(this.element.style, true);
-    this.parentElement.appendChild(this.element);
-  }
-
-  uninstall() {
-    if (this.element) {
-      this.parentElement.removeChild(this.element);
-      this.parentElement = null;
-      this.element = null;
-    }
-  }
-
-  setMeasureNodeStyles(style, isRoot) {
-    style.width = style.height = "auto";
-    style.left = style.top = "0px";
-    style.visibility = "hidden";
-    style.position = "absolute";
-    style.whiteSpace = "pre";
-    style.font = "inherit";
-    style.overflow = isRoot ? "hidden" : "visible";
-  }
-
-  measure(fontFamily, fontSize, char) {
-    var rect = null;
-    this.element.innerHTML = char;
-    this.element.style.fontFamily = fontFamily;
-    this.element.style.fontSize = fontSize + "px";
-    try {
-      rect = this.element.getBoundingClientRect();
-    } catch(e) {
-      rect = {width: 0, height:0};
-    };
-    return {
-      height: rect.height,
-      width: rect.width
-    }
-  }
-
-  sizeFor(fontFamily, fontSize, char) {
-    if (char.length > 1)
-      return this.sizeForStr(fontFamily, fontSize, char);
-
-    if (!this.charMap[fontFamily]) {
-      this.charMap[fontFamily] = [];
-    }
-    if (!this.charMap[fontFamily][fontSize]) {
-      this.charMap[fontFamily][fontSize] = [];
-    }
-    if (!this.charMap[fontFamily][fontSize][char])
-      this.charMap[fontFamily][fontSize][char] = this.measure(fontFamily, fontSize, char);
-    return this.charMap[fontFamily][fontSize][char];
-  }
-
-  sizeForStr(fontFamily, fontSize, str) {
-    var height = 0, width = 0;
-    for (let line of str.split('\n')) {
-      let lineHeight = 0, lineWidth = 0;
-      for (let char of line.split('')) {
-        let { height: charHeight, width: charWidth } = this.sizeFor(fontFamily, fontSize, char);
-        if (charHeight > lineHeight) lineHeight = charHeight;
-        lineWidth += charWidth;
-      }
-      if (lineWidth > width) width = lineWidth;
-      height += lineHeight || this.sizeFor(fontFamily, fontSize, " ").height;
-    }
-    return { height: height, width: width };
-  }
-
-  asciiSizes(fontFamily, fontSize) {
-    var result = {};
-    for (var i = 32; i <= 126; i++) {
-      var char = String.fromCharCode(i);
-      result[char] = this.sizeFor(fontFamily, fontSize, char)
-    }
-    return result;
-  }
 }
